requirements/base.txt
fastmcp==2.14.5
neo4j==5.25.0
python-dotenv==1.0.1
